chore(items): break out item methods into individual files
for legibility

AFFECTS PACKAGES:
@esri/arcgis-rest-items
