+++ packages/arcgis-rest-items/src/get.ts
@@ -1,0 +1,76 @@
+import {
+  request,
+  IRequestOptions,
+  getPortalUrl
+} from "@esri/arcgis-rest-request";
+
+import { IItem } from "@esri/arcgis-rest-common-types";
+
+import { IItemIdRequestOptions, IItemDataRequestOptions } from "./helpers";
+
+/**
+ * Get an item by id
+ *
+ * @param id - Item Id
+ * @param requestOptions - Options for the request
+ * @returns A Promise that will resolve with the data from the response.
+ */
+export function getItem(
+  id: string,
+  requestOptions?: IRequestOptions
+): Promise<IItem> {
+  const url = `${getPortalUrl(requestOptions)}/content/items/${id}`;
+
+  // default to a GET request
+  const options: IRequestOptions = {
+    ...{ httpMethod: "GET" },
+    ...requestOptions
+  };
+  return request(url, options);
+}
+
+/**
+ * Get the /data for an item.
+ * @param id - Item Id
+ * @param requestOptions - Options for the request
+ * @returns A Promise that will resolve with the json data for the item.
+ */
+export function getItemData(
+  id: string,
+  requestOptions?: IItemDataRequestOptions
+): Promise<any> {
+  const url = `${getPortalUrl(requestOptions)}/content/items/${id}/data`;
+  // default to a GET request
+  const options: IItemDataRequestOptions = {
+    ...{ httpMethod: "GET", params: {} },
+    ...requestOptions
+  };
+
+  if (options.file) {
+    options.params.f = null;
+  }
+
+  return request(url, options);
+}
+
+/**
+ * Get the resources associated with an item
+ *
+ * @param requestOptions - Options for the request
+ * @returns A Promise to get some item resources.
+ */
+export function getItemResources(
+  requestOptions: IItemIdRequestOptions
+): Promise<any> {
+  const url = `${getPortalUrl(requestOptions)}/content/items/${
+    requestOptions.id
+  }/resources`;
+
+  // mix in user supplied params
+  requestOptions.params = {
+    ...requestOptions.params,
+    num: 1000
+  };
+
+  return request(url, requestOptions);
+}
