+++ packages/arcgis-rest-items/src/index.ts
@@ -1,1 +1,8 @@
-export * from "./items";
+export * from "./add";
+export * from "./create";
+export * from "./get";
+export * from "./protect";
+export * from "./remove";
+export * from "./search";
+export * from "./update";
+export * from "./helpers";
