+++ packages/arcgis-rest-items/src/remove.ts
@@ -1,0 +1,55 @@
+import { request, getPortalUrl } from "@esri/arcgis-rest-request";
+
+import {
+  IItemIdRequestOptions,
+  IItemResourceRequestOptions,
+  determineOwner
+} from "./helpers";
+
+/**
+ * Remove an item from the portal
+ *
+ * *
+ * ```js
+ * import { removeItem } from '@esri/arcgis-rest-items';
+ *
+ * removeItem({
+ *   authentication: userSession,
+ *   id: "3ef"
+ * })
+ * ```
+ *
+ * @param requestOptions - Options for the request
+ * @returns A Promise that deletes an item.
+ */
+export function removeItem(
+  requestOptions: IItemIdRequestOptions
+): Promise<any> {
+  const owner = determineOwner(requestOptions);
+  const url = `${getPortalUrl(requestOptions)}/content/users/${owner}/items/${
+    requestOptions.id
+  }/delete`;
+  return request(url, requestOptions);
+}
+
+/**
+ * Remove a resource associated with an item
+ *
+ * @param requestOptions - Options for the request
+ * @returns A Promise to unprotect an item.
+ */
+export function removeItemResource(
+  requestOptions: IItemResourceRequestOptions
+): Promise<any> {
+  const owner = determineOwner(requestOptions);
+  const url = `${getPortalUrl(requestOptions)}/content/users/${owner}/items/${
+    requestOptions.id
+  }/removeResources`;
+
+  // mix in user supplied params
+  requestOptions.params = {
+    ...requestOptions.params,
+    resource: requestOptions.resource
+  };
+  return request(url, requestOptions);
+}
