+++ packages/arcgis-rest-items/src/search.ts
@@ -1,0 +1,70 @@
+import {
+  request,
+  IRequestOptions,
+  getPortalUrl
+} from "@esri/arcgis-rest-request";
+
+import { IPagingParams, IItem } from "@esri/arcgis-rest-common-types";
+
+// this interface still needs to be docced
+export interface ISearchRequest extends IPagingParams {
+  q: string;
+  [key: string]: any;
+}
+
+export interface ISearchRequestOptions extends IRequestOptions {
+  searchForm?: ISearchRequest;
+}
+
+/**
+ * Options to pass through when searching for items.
+ */
+export interface ISearchResult {
+  query: string; // matches the api's form param
+  total: number;
+  start: number;
+  num: number;
+  nextStart: number;
+  results: IItem[];
+}
+
+/**
+ * Search for items via the portal api
+ *
+ * ```js
+ * import { searchItems } from '@esri/arcgis-rest-items';
+ *
+ * searchItems('water')
+ * .then((results) => {
+ *  console.log(results.total); // 355
+ * })
+ * ```
+ *
+ * @param search - A string or RequestOptions object to pass through to the endpoint.
+ * @returns A Promise that will resolve with the data from the response.
+ */
+export function searchItems(
+  search: string | ISearchRequestOptions
+): Promise<ISearchResult> {
+  let options: ISearchRequestOptions = {
+    httpMethod: "GET",
+    params: {}
+  };
+
+  if (typeof search === "string") {
+    options.params.q = search;
+  } else {
+    options.params = search.searchForm;
+    // mixin, giving user supplied requestOptions precedence
+    options = {
+      ...options,
+      ...search
+    };
+  }
+
+  // construct the search url
+  const url = `${getPortalUrl(options)}/search`;
+
+  // send the request
+  return request(url, options);
+}
