+++ packages/arcgis-rest-items/test/get.test.ts
@@ -1,0 +1,178 @@
+import * as fetchMock from "fetch-mock";
+
+import { getItem, getItemData, getItemResources } from "../src/get";
+
+import { ItemResponse, ItemDataResponse } from "./mocks/item";
+
+import { GetItemResourcesResponse } from "./mocks/resources";
+
+import { UserSession } from "@esri/arcgis-rest-auth";
+import { TOMORROW } from "@esri/arcgis-rest-auth/test/utils";
+
+describe("get", () => {
+  afterEach(fetchMock.restore);
+
+  it("should return an item by id", done => {
+    fetchMock.once("*", ItemResponse);
+
+    getItem("3ef")
+      .then(response => {
+        expect(fetchMock.called()).toEqual(true);
+        const [url, options]: [string, RequestInit] = fetchMock.lastCall("*");
+        expect(url).toEqual(
+          "https://www.arcgis.com/sharing/rest/content/items/3ef?f=json"
+        );
+        expect(options.method).toBe("GET");
+        done();
+      })
+      .catch(e => {
+        fail(e);
+      });
+  });
+
+  it("should return an item data by id", done => {
+    fetchMock.once("*", ItemDataResponse);
+
+    getItemData("3ef")
+      .then(response => {
+        expect(fetchMock.called()).toEqual(true);
+        const [url, options]: [string, RequestInit] = fetchMock.lastCall("*");
+        expect(url).toEqual(
+          "https://www.arcgis.com/sharing/rest/content/items/3ef/data?f=json"
+        );
+        expect(options.method).toBe("GET");
+        done();
+      })
+      .catch(e => {
+        fail(e);
+      });
+  });
+
+  it("should return binary item data by id", done => {
+    // Blob() is only available in the browser
+    if (typeof window !== "undefined") {
+      fetchMock.once("*", {
+        sendAsJson: false,
+        headers: { "Content-Type": "application/zip" },
+        body: new Blob()
+      });
+
+      getItemData("3ef", { file: true })
+        .then(response => {
+          expect(fetchMock.called()).toEqual(true);
+          const [url, options]: [string, RequestInit] = fetchMock.lastCall("*");
+          expect(url).toEqual(
+            "https://www.arcgis.com/sharing/rest/content/items/3ef/data"
+          );
+          expect(options.method).toBe("GET");
+          expect(response instanceof Blob).toBeTruthy();
+          done();
+        })
+        .catch(e => {
+          fail(e);
+        });
+    } else {
+      done();
+    }
+  });
+
+  describe("Authenticated methods", () => {
+    // setup a UserSession to use in all these tests
+    const MOCK_USER_SESSION = new UserSession({
+      clientId: "clientId",
+      redirectUri: "https://example-app.com/redirect-uri",
+      token: "fake-token",
+      tokenExpires: TOMORROW,
+      refreshToken: "refreshToken",
+      refreshTokenExpires: TOMORROW,
+      refreshTokenTTL: 1440,
+      username: "casey",
+      password: "123456",
+      portal: "https://myorg.maps.arcgis.com/sharing/rest"
+    });
+
+    const MOCK_USER_REQOPTS = {
+      authentication: MOCK_USER_SESSION
+    };
+
+    it("should return an item by id using a token", done => {
+      fetchMock.once("*", ItemResponse);
+
+      getItem("3ef", MOCK_USER_REQOPTS)
+        .then(response => {
+          expect(fetchMock.called()).toEqual(true);
+          const [url, options]: [string, RequestInit] = fetchMock.lastCall("*");
+          expect(url).toEqual(
+            "https://myorg.maps.arcgis.com/sharing/rest/content/items/3ef?f=json&token=fake-token"
+          );
+          expect(options.method).toBe("GET");
+          done();
+        })
+        .catch(e => {
+          fail(e);
+        });
+    });
+
+    it("should return an item data by id using a token", done => {
+      fetchMock.once("*", ItemDataResponse);
+
+      getItemData("3ef", MOCK_USER_REQOPTS)
+        .then(response => {
+          expect(fetchMock.called()).toEqual(true);
+          const [url, options]: [string, RequestInit] = fetchMock.lastCall("*");
+          expect(url).toEqual(
+            "https://myorg.maps.arcgis.com/sharing/rest/content/items/3ef/data?f=json&token=fake-token"
+          );
+          expect(options.method).toBe("GET");
+          done();
+        })
+        .catch(e => {
+          fail(e);
+        });
+    });
+
+    it("get item resources", done => {
+      fetchMock.once("*", GetItemResourcesResponse);
+      getItemResources({
+        id: "3ef",
+        ...MOCK_USER_REQOPTS
+      })
+        .then(response => {
+          const [url, options]: [string, RequestInit] = fetchMock.lastCall("*");
+          expect(url).toEqual(
+            "https://myorg.maps.arcgis.com/sharing/rest/content/items/3ef/resources"
+          );
+          expect(options.method).toBe("POST");
+          expect(options.body).toContain("f=json");
+          done();
+        })
+        .catch(e => {
+          fail(e);
+        });
+    });
+
+    it("get item resources with extra parameters", done => {
+      fetchMock.once("*", GetItemResourcesResponse);
+      getItemResources({
+        id: "3ef",
+        ...MOCK_USER_REQOPTS,
+        params: {
+          resourcesPrefix: "foolder"
+        }
+      })
+        .then(response => {
+          const [url, options]: [string, RequestInit] = fetchMock.lastCall("*");
+          expect(url).toEqual(
+            "https://myorg.maps.arcgis.com/sharing/rest/content/items/3ef/resources"
+          );
+          expect(options.method).toBe("POST");
+          expect(options.body).toContain("f=json");
+          expect(options.body).toContain("resourcesPrefix=foolder");
+          done();
+        })
+        .catch(e => {
+          fail(e);
+        });
+    });
+  }); // auth requests
+});
